Move geth input dump into another file
